Move protobuf logic from buffer crate to language crate
This will enable us to add operations that only pertain to the language crate.
